+++ scout/models/case/case_loading_models.py
@@ -17,7 +17,11 @@ except ImportError:
 
 from pydantic import BaseModel, Field, field_validator, model_validator
 
-from scout.constants import ANALYSIS_TYPES, FILE_TYPE_MAP, OMICS_FILE_TYPE_MAP
+from scout.constants import (
+    ANALYSIS_TYPES,
+    ORDERED_FILE_TYPE_MAP,
+    ORDERED_OMICS_FILE_TYPE_MAP,
+)
 from scout.exceptions import PedigreeError
 from scout.utils.date import get_date
 
@@ -63,8 +67,8 @@ CASE_FILE_PATH_CHECKS = [
     "rna_delivery_report",
 ]
 
-VCF_FILE_PATH_CHECKS = FILE_TYPE_MAP.keys()
-OMICS_FILE_PATH_CHECKS = OMICS_FILE_TYPE_MAP.keys()
+VCF_FILE_PATH_CHECKS = ORDERED_FILE_TYPE_MAP.keys()
+OMICS_FILE_PATH_CHECKS = ORDERED_OMICS_FILE_TYPE_MAP.keys()
 
 GENOME_BUILDS = ["37", "38"]
 TRACKS = ["rare", "cancer"]